Move module from references to script
